installer done. documented code.

+++ Backend/install.py
@@ -4,8 +4,7 @@ import sys
 import _winreg
 import errno
 import ctypes
-import subprocess
-from time import sleep
+import time
 
 FILE_ATTRIBUTE_HIDDEN = 0x02
 
@@ -17,50 +16,98 @@ else:
 	print "Platform Not Supported"
 	print "Multi-platform support will be added soon"
 	print "Contact us for further information"
-	sleep(10)
+	time.sleep(10)
 	sys.exit(0)
+	
+try:
+	os.makedirs(system_folder)
+except OSError as e:
+	if e.errno == 5 or e.errno==13:
+		print "Installation Failed 1"
+		print "Please run as an ADMINISTRATOR"
+		time.sleep(10)
+		sys.exit(0)
+	elif e.errno==17:
+		try:
+			print "You have droplet previously installed. Clearing up in 5 seconds"
+			time.sleep(2)
+			os.startfile(os.path.join(os.getcwd(),"killall.exe"))
+			print "All ready. Installing now..."
+			time.sleep(3)
+			shutil.rmtree(system_folder)
+			os.makedirs(system_folder)
+		except OSError as newe:
+			if newe.errno == 5 or newe.errno==13:
+				print "Installation Failed 1"
+				print "Please run as an ADMINISTRATOR"
+				time.sleep(10)
+				sys.exit(0)
+			print "Error. Installation Failed 1"
+			print "RE-RUN THE INSTALLER. If error persists please contact an administrator and say ERROR NUMBER 1 and error "+str(newe.errno)+". It is very important to correct this error now and in further releases"
+			time.sleep(10)
+			sys.exit(0)
+	else:
+		print "Error. Installation Failed 1"
+		print "RE-RUN THE INSTALLER. If error persists please contact an administrator and say ERROR NUMBER 2 and error "+str(e.errno)+". It is very important to correct this error now and in further releases"
+		time.sleep(10)
+		sys.exit(0)
+flag=1	
 try:
 	os.makedirs(path)
 except OSError as e:
-	print type(e.errno)
 	if e.errno==5 or e.errno==13:
-		print "Installation Failed 1"
+		print "Installation Failed 2"
 		print "Please run as an ADMINISTRATOR"
-		sleep(10)
+	elif e.errno==17:
+		flag=0
+	else:
+		print "Error. Installation Failed 2"
+		print "RE-RUN THE INSTALLER. If error persists please contact an administrator and say ERROR NUMBER 3 and error "+str(e.errno)+". It is very important to correct this error now and in further releases"
+	if flag==1:
+		print "Cleaning up..."
+		try:
+			shutil.rmtree(system_folder)
+			print "Done. State : 1RESET"
+		except:
+			print "Problem during clean-up 1. Run as an ADMINISTRATOR. If problem persists contact us."
+		time.sleep(10)
 		sys.exit(0)
-		pass
+		
 try:
 	os.makedirs(config)#By this point you are already admin but just to be safe... and lazy.
 except OSError as e:
+	flag2=1
 	if e.errno ==5 or e.errno==13:
-		print "Installation Failed 2"
+		print "Installation Failed 3"
 		print "Please run as an ADMINISTRATOR"
-		sleep(10)
-		sys.exit(0)
-		pass
+	elif e.errno==17:
+		try:
+			shutil.rmtree(config)
+			os.makedirs(config)
+			flag2=0
+		except OSError as newe:
+			print "Error. Installation Failed 3"
+			print "RE-RUN THE INSTALLER. If error persists please contact an administrator and say ERROR NUMBER 4 and error "+str(newe.errno)+". It is very important to correct this error now and in further releases"
 	else:
-		print "Error. "
-		print "Please contact an administrator and say ERROR NUMBER 3 and error "+str(e.errno)+". It is very important to correct this error now and in further releases"
-		sleep(1)
+		print "Error. Installation Failed 3"
+		print "RE-RUN THE INSTALLER. If error persists please contact an administrator and say ERROR NUMBER 5 and error "+str(e.errno)+". It is very important to correct this error now and in further releases"
+	if flag2==1:
+		print "Cleaning up..."
+		try:
+			shutil.rmtree(system_folder)
+			if flag==1:
+				shutil.rmtree(path)
+				print "Done. State : 2RESET"
+			print "3RESET"
+		except:
+			print "Problem during clean-up 2. Run as an ADMINISTRATOR. If problem persists contact us."
+		time.sleep(10)
+		sys.exit(0)
+
 ret = ctypes.windll.kernel32.SetFileAttributesW(ur"C:\\droplet\\config",FILE_ATTRIBUTE_HIDDEN)
 if ret ==0:
-	raise ctypes.WinError()
-if not os.path.isdir(system_folder):
-	try:
-		os.makedirs(system_folder)
-	except OSError as e:
-		if e.errno == 5 or e.errno==13:
-			print "Installation Failed 3"
-			print "Please run as an ADMINISTRATOR"
-			pass
-		else:
-			print "Error. Installation Failed"
-			print "Please contact an administrator and say ERROR NUMBER 4 and error "+str(e.errno)+". It is very important to correct this error now and in further releases"
-		print "Cleaning up..."
-		shutil.rmtree(path)
-		print "Done."
-		sleep(10)
-		sys.exit(0)
+	raise ctypes.WinError() #And then what?
+
 files=[ f for f in os.listdir(os.getcwd()) if os.path.isfile(os.path.join(os.getcwd(),f))]
 for f in files:
 	if f != "networks":
@@ -69,20 +116,23 @@ for f in files:
 		except IOError as e:
 			if e.errno==13 or e.errno==5:
 				print "Installation Failed 4"
-				print "Please run as an ADMINSTRATOR"
+				print "Please run as an ADMINISTRATOR"
 			else:
-				print "Error. Installation Failed"
-				print "Please contact an administrator and say ERROR NUMBER 5 and error "+str(e.errno)+". It is very important to correct this error now and in further releases"
-			print "Cleaning up..."
-			shutil.rmtree(path)
-			print "Done."
-			sleep(10)
+				print "Error. Installation Failed 4"
+				print "RE-RUN THE INSTALLER. If error persists please contact an administrator and say ERROR NUMBER 6 and error "+str(e.errno)+" with file "+e.filename+" . It is very important to correct this error now and in further releases"
+				print "Cleaning up..."
+			try:
+				shutil.rmtree(system_folder)
+				if flag==1:
+					shutil.rmtree(path)
+					print "Done. State : 4RESET"
+				print "5RESET"
+			except:
+				print "Problem during clean-up 3. Run as an ADMINISTRATOR. If problem persists contact us."
+			time.sleep(10)
 			sys.exit(0)
 			
 shutil.copy("networks",config)
-#shutil.copy("upload.py",system_folder)
-#shutil.copy("hola.py",system_folder)
-#shutil.copy("run.exe",system_folder)
 
 _winreg.CreateKey(_winreg.HKEY_CLASSES_ROOT, 'drop\\shell\\open\\command')
 _winreg.CreateKey(_winreg.HKEY_CLASSES_ROOT,'drop\\DefaultIcon')
@@ -94,13 +144,9 @@ stkey= _winreg.OpenKey(_winreg.HKEY_LOCAL_MACHINE,"SOFTWARE\\Microsoft\\Windows\
 _winreg.SetValueEx(stkey,"Droplet_sender",0,_winreg.REG_SZ,'"'+system_folder+'sender.exe"')
 _winreg.SetValueEx(stkey,"Droplet_hashing",0,_winreg.REG_SZ,'"'+system_folder+'hashing.exe"')
 
-#os.system("bckghashing.vbs")###subprocess is better
-#os.system("bckgsender.vbs")###subprocess is better
+shutil.copy("droplet.exe",os.environ['appdata']+r"\Microsoft\Windows\Start Menu\Programs\Startup")
+os.startfile(system_folder+"droplet.exe")
 
-###REmove and run
-os.system("start /b \"\" \"C:\\Program Files (x86)\\droplet\\pickhash.exe\"")
-os.system("start /b \"\" \"C:\\Program Files (x86)\\droplet\\sender.exe\"")
-shutil.copy("exp.bat",os.environ['appdata']+r"\Microsoft\Windows\Start Menu\Programs\Startup")
-print "Installation Success"
+print "\n\nInstallation Success!!"
 print "Start downloading files from the website! Please remember to allow firewall access to the programs when they run for the first time"
-sleep(10)
+time.sleep(10)
